--- a/packages/turf-buffer/package.json
+++ b/packages/turf-buffer/package.json
@@ -32,6 +32,7 @@
   },
   "homepage": "https://github.com/Turfjs/turf",
   "devDependencies": {
+    "@turf/helpers": "^4.0.1",
     "@turf/truncate": "^4.0.1",
     "benchmark": "^2.1.4",
     "load-json-file": "^2.0.0",
@@ -39,9 +40,12 @@
     "write-json-file": "^2.0.0"
   },
   "dependencies": {
-    "@mapbox/geojson-normalize": "0.0.1",
-    "@turf/combine": "^4.0.2",
+    "@mapbox/geojson-normalize": "^0.0.1",
+    "@turf/circle": "^4.0.1",
     "@turf/helpers": "^4.0.1",
-    "jsts": "1.3.0"
+    "@turf/invariant": "^4.0.1",
+    "@turf/meta": "^4.0.2",
+    "jsts": "^1.3.0",
+    "polygon-offset": "^0.3.1"
   }
 }