Attempting to replace `jsts` from `@turf/buffer`, my findings are good & bad:

## Good
- Using [`@turf/circle`](https://github.com/Turfjs/turf/tree/master/packages/turf-circle) to create `(Multi)Points` buffers is 2-4x faster and has very minimal source code.

## Bad
- Handling (Multi)Polygons is 8-12x SLOWER using [`polygon-offset`](https://github.com/w8r/polygon-offset).
- the [`martinez`] library has a hard time determining the differences between `Polygon` & `MultiPolygon`, the output coordinates are always formatted for a Polygon.
- Edges of new buffered polygons are jagged (`jsts` has smooth edges)
![image](https://cloud.githubusercontent.com/assets/550895/24847154/2f024df2-1d8d-11e7-9970-8f583e02f0aa.png)

## Benchmark results

```
==before==
feature-collection-points x 3,792 ops/sec ±10.41% (82 runs sampled)
geometry-collection-points x 4,346 ops/sec ±2.05% (90 runs sampled)
linestring x 9,087 ops/sec ±2.14% (89 runs sampled)
multi-linestring x 1,145 ops/sec ±9.46% (80 runs sampled)
multi-point x 4,898 ops/sec ±4.73% (78 runs sampled)
multi-polygon x 1,737 ops/sec ±9.31% (66 runs sampled)
point x 11,907 ops/sec ±8.10% (72 runs sampled)
polygon-with-holes x 6,417 ops/sec ±6.16% (79 runs sampled)

==after==
feature-collection-points x 6,198 ops/sec ±13.12% (58 runs sampled)
geometry-collection-points x 6,518 ops/sec ±13.37% (67 runs sampled)
linestring x 1,694 ops/sec ±4.85% (71 runs sampled)
multi-linestring x 466 ops/sec ±4.26% (77 runs sampled)
multi-point x 12,055 ops/sec ±5.63% (77 runs sampled)
multi-polygon x 202 ops/sec ±1.73% (83 runs sampled)
point x 43,109 ops/sec ±1.58% (89 runs sampled)
polygon-with-holes x 438 ops/sec ±1.70% (86 runs sampled)
```

## Summary

So far, looks like `jsts` is here to stay until we find a better replacement solution.

CC: @morganherlocker @rowanwins