@rowanwins 2017-04-11
Thanks for posting the results of your investigation @DenisCarriere , a bit disappointing but let's not lose all hope just yet. I suggest we a look upstream in poylgon-offset and see if the is anything we can contribute the that may aid the cause