@DenisCarriere 2017-04-16
Going to close this PR, the `martinez` library doesn't seem like a solid replacement for `jsts` (at the moment).