@rowanwins 2017-04-16
Gday @DenisCarriere 

I'd leave the pull request open while we explore things just so it's here for reference.

Another thing I noticed in the `polygon-offset` library is that there is an `arcSegments` option that you can pass through. It [defaults to 5](https://github.com/w8r/polygon-offset/blob/master/src/offset.js#L49) (I suspect as we're seeing in your image above). I guess you could use this to generate different corner types (eg a mitre corner vs a curved corner) which may actually be handy.

Still need to look at the performance thing further (perhaps it's worth trying and seeing what happens speed wise if you set arcSegments = 1).