@DenisCarriere 2017-04-16
Agreed, however all the no-breaking changes from this PR was translated to: https://github.com/Turfjs/turf/pull/667.

I will be opening a new `replace jsts for turf-buffer` PR starting from the most recent `master` branch.

👍 Yes I didn't know about those `arcSegments`, using `64 steps` looks the same as `jsts`, unfortunately the performance is still significantly slower than `jsts`.